cavs: shim: make register functions common
Moves register helper functions to common cavs header.
